+++ cmd/root.go
@@ -3,8 +3,8 @@ package cmd
 import (
 	"fmt"
 	"github.com/spf13/viper"
+	"github.krafton.com/sbx/oidc-discovery-server/issuer_provider"
 	"github.krafton.com/sbx/oidc-discovery-server/server"
-	"github.krafton.com/sbx/oidc-discovery-server/server/issuer_provider"
 	"go.uber.org/zap"
 	"net/http"
 	"net/url"
