+++ server/key_provider.go
@@ -6,8 +6,8 @@ import (
 	cmap "github.com/orcaman/concurrent-map/v2"
 	"github.com/pkg/errors"
 	"github.com/zitadel/oidc/v2/pkg/op"
+	"github.krafton.com/sbx/oidc-discovery-server/issuer_provider"
 	"github.krafton.com/sbx/oidc-discovery-server/jwt"
-	"github.krafton.com/sbx/oidc-discovery-server/server/issuer_provider"
 	"github.krafton.com/sbx/oidc-discovery-server/util"
 	"go.uber.org/zap"
 	"net/http"
